% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat_taxonomy_matrix.R
\docType{package}
\name{as_flat_taxonomy_matrix}
\alias{as_flat_taxonomy_matrix}
\title{Flatten a taxonomy and encode it as a matrix with one column per taxonomic level}
\usage{
as_flat_taxonomy_matrix(x, id, parent_id, .x = rlang::caller_arg(x))
}
\arguments{
\item{x}{An object of class \code{phylo} (e.g. result of \link[ape:read.tree]{ape::read.tree}) or a data frame with
columns \code{id} and \code{parent_id} (adjacency table).}

\item{id}{Character vector of node identifiers. This cannot be supplied at the same time as \code{x}.}

\item{parent_id}{Character vector of node parent identifiers. This cannot be supplied at the same
time as \code{x}.}

\item{.x}{Optional label for the argument \code{x} to use in error messages}
}
\value{
A data frame with one row per node id and as many columns as there are levels in the tree.
}
\description{
This function converts a taxonomy tree (a phylo object or an adjacency table) into a flat
representation. This is used internally by \link{densify} to traverse the taxonomy tree. The
package user does not need to invoke this function directly, but we chose to expose it since it
can be useful for other applications.
}
\details{
The input taxonomy can be provided either as a \code{phylo} object (e.g. via \link[ape:read.nexus]{ape::read.nexus}), as a
data frame that contains columns \code{id} and \code{parent_id} (each row encodes one parent-child
relationship), or by passing in \code{id} and \code{parent_id} vectors explicitly. The vectors \code{id} and
\code{parent_id} must be of type character.

A taxonomy tree is said to be "flattened" if the depth of each of its leaves is equal to the
tree height. Such a tree can be constructed by injecting dummy nodes where appropriate. Example:

\preformatted{
   A                              A      F''     level 1
  / \         dummy nodes        / \     |
  B  \       ------------->      B  E'   F'      level 2
 / \  \                         / \  \   |
C   D  E  F                    C   D  E  F       leaf level
}

A flat taxonomy can be trivially encoded in a tabular format as it has the full set of nodes at every
level. Each column encodes a taxonomic level. The main purpose of this representation is the
ability to find the parent of every leaf at every level in O(1).

Note that there can be multiple flattened representations for a given tree, since it is not
guaranteed that nodes at different levels can be aligned in the unique way. In our
implementation we always align at the top level, which means that the dummy nodes are always
inserted at the bottom, between the leaf and it's closest parent in the original tree.
}
\examples{
as_flat_taxonomy_matrix(glottolog_languoids)

}
