% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_flat_taxonomy_matrix.R
\docType{package}
\name{build_flat_taxonomy_matrix}
\alias{build_flat_taxonomy_matrix}
\title{Flatten input taxonomy}
\usage{
build_flat_taxonomy_matrix(id, parent_id)
}
\arguments{
\item{id}{Vector of node ids.}

\item{parent_id}{Vector of node parent ids (must be of same size and type as \code{id}).}
}
\value{
A data frame with one row per node id and as many columns as there are levels in the tree.
}
\description{
Flattens a taxonomy and encodes it in a table (with one column per taxonomic level)
}
\details{
A taxonomy tree is said to be "flattened" if the depth of each of it's leaf nodes is equal to the tree height. Such tree can be constructed by injecting dummy nodes where appropriate. Example:

\preformatted{   A                              A      F''     level 1
  / \         dummy nodes        / \     |
  B  \       ------------->      B  E'   F'      level 2
 / \  \                         / \  \   |
 C  D  E  F                     C  D  E  F       leaf level}

A flat taxonomy can be trivially encoded in a tabular format as it has nodes at every level. Each
column encodes a taxonomic level. The main purpose of this representation is to query a leaf
parent at level k in O(1).

Note that there can be multiple flattened representations for a given tree, since there it is not
guaranteed that nodes at different levels can be aligned in the unique way. In our implementation
we always align at the top level, which means that the dummy nodes are always inserted at the
bottom, between the leaf and it's closest parent in the original tree.
}
\examples{
build_flat_taxonomy_matrix(glottolog_languoids$id, glottolog_languoids$parent_id)

}
