% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify.R
\name{densify}
\alias{densify}
\title{Iterative matrix densification}
\usage{
densify(
  data,
  cols,
  ...,
  taxonomy,
  taxon_id,
  scoring = c("log_odds", "arithmetic", "geometric"),
  min_variability = 1,
  limits = list(min_coding_density = 1, min_prop_rows = NA, min_prop_cols = NA),
  scoring_weights = list(coding = 1, taxonomy = 1)
)
}
\arguments{
\item{data}{A data frame with observations/taxa (such as languages) in rows and variables in columns.}

\item{cols}{<\code{\link[tidyselect:language]{tidy-select}}> specification of variable columns to densify (default is all columns are treated as variable
columns). Columns not specified here will be preserved during densification.}

\item{taxonomy}{A taxonomy tree, which can be a \code{phylo} object (e.g. result of \link[ape:read.tree]{ape::read.tree} or a data frame with columns \code{id} and \code{parent_id}
(such as \link{glottolog_languoids}). A taxonomy must be provided as a named argument if you want to consider the taxonomic diversity for densification.}

\item{taxon_id}{The name of the column with taxa identifiers. The \code{data} must contain such a column if a taxonomy is supplied. If not supplied,
the function will try to make an educated guess based on the column contents.}

\item{scoring}{A character string specifying the type of scoring used for calculating the importance weights. Possible values are are "arithmetic",
"geometric", or "log_odds". Default is "log_odds".}

\item{min_variability}{An integer specifying the minimal threshold of the second-most-frequent state of any variable. Variables below this threshold
will be discarded. Supply \code{NA} to disable variability pruning. Default is \code{1}.}

\item{limits}{A named list specifying the limit conditions for the densification process termination. Available limit conditions are
\code{min_coding_density} (a number between 0 and 1 specifying the target coding density), \code{min_prop_rows} (a number between 0 and 1 specifying the
minimal proportion of rows that have to be retained in the data), and \code{min_prop_cols} (a number between 0 and 1 specifying the
minimal proportion of variable columns that have to be retained in the data). More then one condition can be specified simultaneously, \code{densify()}
will stop if any condition is reached.}

\item{scoring_weights}{A named list specifying additional weighting factors to be applied during importance score calculation. Supported parameters
are \code{coding} and \code{taxonomy}, which apply to coding and taxonomic importance weights, respectively. The values for these parameters must be a
number between 0 and 1. Specify 0 (or \code{NULL}, or \code{NA}) to disable the respective weight calculation. Use these parameters to tweak the relative
importance between the available coding and taxonomic diversity. Note: the weighting factor is applied before the scoring function, experimentation
will be required to find a good balance of parameters for \code{scoring = "log_odds"}. See the vignette for more details.}
}
\value{
A specially formatted tibble of iterative densification results, \link{densify_result}
}
\description{
Iteratively densifies an input data frame, returning a specially formatted tible (\link{densify_result}) describing the result of each densification step.
Densified data can be retrieved using function \link[=prune.densify_result]{prune}, or manually by invoking \link{as_tibble} on the values in the column \code{data}. The densification results
can also be inspected using \link[=visualize.densify_result]{visualize}
}
\examples{
densify(WALS[1:100, 1:100], taxonomy = glottolog_languoids)

}
