% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify.R
\name{densify_prep}
\alias{densify_prep}
\title{pre-processing linguistic input data frame}
\usage{
densify_prep(OriginalData)
}
\arguments{
\item{OriginalData}{input data frame}
}
\value{
prepped matrix
}
\description{
This function pre-processes a linguistic input data frame by replacing NAs with 0
(indicating no data available) and non-NA entries with 1 (indicating data available).
}
