% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_prune.R
\name{densify_prune}
\alias{densify_prune}
\title{Pruning of input matrix, given iteration log and optimum quality score}
\usage{
densify_prune(original_data, iteration_log, optimum = 1)
}
\arguments{
\item{original_data}{A data frame with the observation names as row names and variable names as column names.}

\item{iteration_log}{A data frame containing the iteration_log generated by \code{densify_steps}.}

\item{optimum}{The iteration number representing the optimum quality score determined by \code{densify_score}. The function will densify the matrix and prune rows and columns up to this iteration. Default is set to 1 (output is original data frame).}
}
\value{
A pruned and densified data frame.
}
\description{
Densifies the (original) input data matrix given the iteration log and the optimum quality score retrieved from the \code{densify_steps} and \code{densify_score} functions by pruning the appropriate rows and columns.
}
\examples{
# Assuming 'original_data' and 'iteration_log' are prepared
# pruned_densified <- densify_prune(original_data, iteration_log, optimum = 2)
}
\seealso{
\code{\link{densify_steps}}

\code{\link{densify_score}}
}
