% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify.R
\name{densify_result}
\alias{densify_result}
\title{Result of densify function}
\format{
A tibble with the following columns:
\describe{
\item{.step}{densify process iteration step}
\item{n_data_points}{number of total non-missing data points in the pruned data}
\item{n_rows}{number of rows (taxa) in the pruned data}
\item{n_cols}{numbr of variables in the pruned data (note: number refers specifically to variables, extra columns are always preserved)}
\item{coding_density}{proportion of non-missing data points relative to the data size}
\item{row_coding_density_min}{minimum proportion of non-missing data points, considering each row}
\item{row_coding_density_avg}{mean proportion of non-missing data points, considering each row}
\item{row_coding_density_max}{maximum proportion of non-missing data points, considering each row}
\item{col_coding_density_min}{minimum proportion of non-missing data points, considering each column}
\item{col_coding_density_avg}{mean proportion of non-missing data points, considering each column}
\item{col_coding_density_max}{maximum proportion of non-missing data points, considering each column}
\item{taxonomic_index}{Shannon diversity index for top-level taxa (if taxonomy has been supplied)}
\item{data}{a list containing promise objects that evaluate to pruned datasets (use \link{as.data.frame} to evaluate a promise)}
\item{changes}{a tibble containing information about the pruned data dimensions}
}
}
\description{
A tibble that contains the information about each pruning step performed by \link{densify} along with some useful
statistics. Use \link[=rank_results.densify_result]{rank_results} to rank the densify results according to a quality score, \link[=prune.densify_result]{prune} to retrieve a
subjectively best result, and \link[=visualize.densify_result]{visualize} to visually compare the quality scores between different pruning steps.
You can also do custom evaluation on the results tibble, e.g. by using \code{dplyr} and retrieve the pruned datasets from the
column \code{data}.
}
