% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_score.R
\name{densify_score}
\alias{densify_score}
\title{Quality score computation for each iteration}
\usage{
densify_score(
  iteration_log,
  exponent_prop_coded_data = 1,
  exponent_available_data_points = 1,
  exponent_lowest_taxon_coding_score = NULL,
  exponent_lowest_variable_coding_score = NULL,
  exponent_taxonomic_diversity = NULL,
  plot = FALSE
)
}
\arguments{
\item{iteration_log}{A data frame containing the output iteration_log from the \code{densify_steps} function.}

\item{exponent_prop_coded_data}{Exponent for the proportion of coded data. Default is 1.}

\item{exponent_available_data_points}{Exponent for the number of available data points. Default is 1.}

\item{exponent_lowest_taxon_coding_score}{Exponent for the lowest taxon score. This exponent is not mandatory, so there is no default.}

\item{exponent_lowest_variable_coding_score}{Exponent for the lowest variable score. This exponent is not mandatory, so there is no default.}

\item{exponent_taxonomic_diversity}{Exponent for the taxonomic diversity index. This exponent is not mandatory, so there is no default.}

\item{plot}{Logical denoting whether the quality scores should be plotted. Default is \code{FALSE}.}
}
\value{
The optimal number of iterations with the highest quality score.
}
\description{
Defines a quality score for each iteration as documented in the output created in the \code{densify_steps} function and identifies the optimal number of iterations post-hoc according to modifiable exponents.
}
\examples{
# Assuming `iteration_log` is a data frame generated from `densify_steps` function
# densify_score(iteration_log)

}
\seealso{
\code{\link{densify_steps}}
}
