% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify_steps.R
\name{densify_steps}
\alias{densify_steps}
\title{Iterative matrix densification according to specified criteria}
\usage{
densify_steps(
  original_data = wals,
  max_steps = 1,
  variability_threshold = 1,
  mean_type = "log_odds",
  taxonomy = F,
  taxonomy_matrix,
  tax_weight_factor = 0.99,
  coding_weight_factor = 0.99,
  verbose = FALSE
)
}
\arguments{
\item{original_data}{A data frame with observations/taxa (e.g. languages, as glottocodes) as row names and variable names as column names.
Any question marks, empty entries, or "NA"s must be represented as NA. Default is the data from WALS.}

\item{max_steps}{An integer specifying the maximum number of iterations attempted during densification.
Default is 1.}

\item{variability_threshold}{specifying how many taxa the second-most-frequent variable state of any variable must contain for the variable to be maintained in the matrix.
Default is 1.}

\item{mean_type}{A character string specifying the type of mean to be used for calculating the final weights.
Possible values are "arithmetic", "geometric", or "log_odds". Default is "log_odds".}

\item{taxonomy}{A logical indicating whether taxonomic diversity should be factored in for the pruning/retention of rows.
If \code{TRUE}, taxonomic diversity is considered; if \code{FALSE}, it is not. Default is \code{FALSE}.}

\item{taxonomy_matrix}{A data frame containing a taxonomy matrix, which can be retrieved from the \code{build_flat_taxonomy_matrix} function or provided directly.
This parameter must be specified if \code{taxonomy = TRUE} and is optional otherwise (if specified, taxonomic index is computed for each iteration).}

\item{tax_weight_factor}{A numeric value between 0 and 1 that determines the relative weight given to taxonomy in taxon pruning.
This parameter must be specified if \code{taxonomy = TRUE} and \code{mean_type = "log_odds"}. Default is 0.99.}

\item{coding_weight_factor}{A numeric value between 0 and 1 that determines the relative weight given to coding quality
(absolute coding score and weighted coding score) in taxon pruning. This parameter must be specified if
\code{mean_type = "log_odds"}. Default is 0.99.}

\item{verbose}{A logical indicating how verbose the function should be. If \code{verbose = FALSE}, only the iteration number is logged. Default is \code{FALSE}.}
}
\value{
An iteration log in the form of a data frame with details about the matrix after each iteration of densification.
}
\description{
Iteratively densifies an input data frame. The output of this densification is a log-file, specifying details about the matrix after each iteration.
}
\examples{
# Assuming `original_data` and `taxonomy_matrix` are appropriate data frames
# densify_steps(original_data = wals,
# max_steps = 3,
# variability_threshold = 3,
# mean_type = "log_odds",
# taxonomy = T,
# taxonomy_matrix,
# tax_weight_factor = 0.99,
# coding_weight_factor = 0.99,
# verbose = T)

}
