% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glottolog_languoids.R
\docType{data}
\name{glottolog_languoids}
\alias{glottolog_languoids}
\alias{temp}
\title{Languoids file from Glottolog, Version 4.8}
\format{
An object of class \code{character} of length 1.
}
\source{
Glottolog, Version 4.8
}
\usage{
data(glottolog_languoids)
}
\description{
This dataset is the languoid.csv file from Glottolog, Version 4.8.
}
\references{
\itemize{
\item Hammarström, Harald & Forkel, Robert & Haspelmath, Martin & Bank, Sebastian. 2023.
Glottolog 4.8.
Leipzig: Max Planck Institute for Evolutionary Anthropology.
https://doi.org/10.5281/zenodo.8131084
(Available online at http://glottolog.org, Accessed on 2023-09-05.)
}
}
\keyword{datasets}
