% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R, R/visualize.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{rank_results}
\alias{prune}
\alias{visualize}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{prune}}, \code{\link[generics]{rank_results}}, \code{\link[generics]{visualize}}}
}}

