% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize.densify_result}
\alias{visualize.densify_result}
\alias{plot.densify_result}
\title{Visualise the distrubution of quality scores across densify results}
\usage{
\method{visualize}{densify_result}(x, ..., score = n_data_points * coding_density)

\method{plot}{densify_result}(x, ..., score = n_data_points * coding_density)
}
\arguments{
\item{x}{a \link{densify_result} object}

\item{score}{an expression that computes a densification quality score. Default score aims to maximise both
the number of data points and the coding density. See \link{densify_result} for available statistics that can
be used to compute a suitable score}
}
\value{
a \code{ggplot2} plot
}
\description{
Produces a \code{ggplot2} visualisation showing the distribution of a quality score across the iteration steps
of the densification process and shows the result with the maximal score. The user can provide their own
quality score formula in \code{score}, which can use any statistics provided in a \link{densify_result} tibble.
}
