% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify-package.R
\docType{package}
\name{densify-package}
\alias{densify-package}
\title{densify: Densifying Input Data Frame}
\description{
This package provides functions to densify data frames with missing values.
}
\author{
\strong{Maintainer}: Anna Graff \email{anna.graff@uzh.ch}

Authors:
\itemize{
  \item Marc Lischka \email{marc.lischka@math.uzh.ch}
  \item Taras Zaharko \email{taras.zakharko@uzh.ch}
}

}
\keyword{internal}
