% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densify.R
\name{densify}
\alias{densify}
\title{Iterative matrix densification}
\usage{
densify(
  data,
  cols,
  ...,
  taxonomy,
  taxon_id,
  scoring = c("log_odds", "arithmetic", "geometric"),
  min_variability = NA,
  consider_taxonomic_diversity
)
}
\arguments{
\item{data}{A data frame with observations/taxa (such as languages) in rows and variables in columns.}

\item{cols}{<\code{\link[tidyselect:language]{tidy-select}}> specification of variable columns to densify (default is all columns are treated as variable
columns). Columns not specified here will be preserved during densification.}

\item{taxonomy}{A taxonomy tree, which can be a \code{phylo} object (e.g. result of \link[ape:read.tree]{ape::read.tree} or a data frame columns \code{id} and \code{parent_id}
(such as \link{glottolog_languoids}). A taxonomy must be provided as a named argument if you want to consider the taxonomic diversity for densification.}

\item{taxon_id}{The name of the column with taxa identifiers. The \code{data} must contain such a colun if a taxonomy is supplied. If not supplied,
the function will try to make an educated guess based on the column contents.}

\item{scoring}{A character strict specifying the type of scoring used for calculating the importance weights. Possible values are are "arithmetic",
"geometric", or "log_odds". Default is "log_odds".}

\item{min_variability}{An integer specifying the minimal threshold of the second-most-frequent state of any variable. Variables below this threshold
will be discarded. Supply \code{NA} to disable variability pruning. Default is \code{NA}.}

\item{consider_taxonomic_diversity}{A logical scalar specifying whether the taxonomic diversity of the taxa should be considered for the
densification process. Defaults to \code{TRUE} if taxonomy is supplied.}
}
\value{
A specially formatted tibble of iterative densification results, \link{densify_result}
}
\description{
Iteratively densifies an input data frame, returning a specially formatted tible (\link{densify_result}) describing the result of each densification step.
Densified data can be retrieved using function \link[=prune.densify_result]{prune}, or manually by invoking \link{as_tibble} on the values in the column \code{data}. The desnification results
can also be inspected using \link[=visualize.densify_result]{visualize}
}
\examples{
densify(WALS[1:100, 1:100], taxonomy = glottolog_languoids)

}
