% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wals.R
\docType{data}
\name{WALS}
\alias{WALS}
\title{WALS language-feature matrix}
\source{
WALS Online
}
\usage{
WALS
}
\description{
This dataset is the language-feature matrix generated from WALS Online (retrieved on 20 October 2023).
}
\examples{
WALS
}
\references{
\itemize{
\item Dryer, Matthew S. & Haspelmath, Martin (eds.) 2013.
WALS Online (v2020.3). Zenodo.
https://doi.org/10.5281/zenodo.7385533
(Available online at https://wals.info, Accessed on 2023-10-20.)
}
}
